backend/requirements.txt
deepagents
fastapi
uvicorn[standard]
sse-starlette
tavily-python
langchain-nvidia-ai-endpoints
langchain-community
python-dotenv
